requirements.txt
attrs>=19.3.0
beautifulsoup4>=4.8.1
cftime>=1.0.4.2
coverage>=6.0
Cython>=0.29.14
dask>=2021.10.0 
future>=0.18.0
GDAL>=3.3.0
importlib-metadata>=0.23
iniconfig>=1.1.1
lisflood-utilities>=0.12.19
llvmlite>=0.37.0
lxml>=4.6.5
more-itertools>=7.2.0
netCDF4>=1.5.3
nine>=1.0.0
numba>=0.54.0
numexpr>=2.7.0
numpy>=1.21.0
packaging>=19.2
pandas>=0.25.1
pathlib2>=2.3.5
pkginfo>=1.5.0.1
pluggy>=0.13.0
py>=1.10.0
pyparsing>=2.4.2
pyproj>=2.4.0
pytest>=6.2.5
pytest-cov>=3.0.0
pytest-mock>=3.6.1
python-dateutil>=2.8.1
pytz>=2019.3
PyYAML>=5.4
six>=1.12.0
soupsieve>=1.9.5
toml>=0.10.0
tomli>=1.2.1
toolz>=0.10.0
xarray>=0.20.2
zipp>=0.6.0
